backend/requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# OAuth2
django-oauth-toolkit>=2.3,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# Production server
gunicorn>=21.0,<22.0

# Environment
python-dotenv>=1.0,<2.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code quality
flake8>=6.1,<7.0
black>=23.12,<24.0
isort>=5.13,<6.0
